Create Pair_Of_Array_Equal_to_1.py
This python file finds the pairs in an array such that their difference equals to 1. 

It firstly creates a dictionary such that each key is element of Array i.e. A[i] and the value in dictionary is the number of occurrences. So if it encouters a new key i.e. new A[i] which dictionary does not already had, it creates a key, value pair of (A[i], 1) where 1 is that it found it for the first time. If it already had that key it gets the value of that key and does the val = val +1.

After filling the dictionary, it iterates over the array. 

Now the interesting thing here is, consider this problem of maths, such that target = 1, and let's assume (x, y) such that x - y = 1. And assume x = A[i]. So the unknown remaining is y. 
To find y, we can do is: 
y = x - target 
or
As in code it is written as: diff = A[i] - target at line 24.

So, we assumed that we already knew one unknown and that was x. We assumed that, at every iteration in A,  x is a confirm one number among the pair (x, y) to form x - y = 1 (Yes! this assumption would be true for one iteration, might not be true for next iteration. Well that is the fun thing about maths)

So, say y is the keys of dictionary. And if in a dictionary it founds y, (meaning at that iteration,  x - 1 is present in the dictionary) And these will be the pair we will get.