Create Reverse_A_String.cpp
It has been asked many times to reverse a string, and many people just go and tried to reverse the whole string. But that's not the thing they ask for, you need to first split the string (parsing) then call on a function of reverse on each single tokens (you obtained from strtok()).